033: fixed footer, added baseball functionality
makes it baseball diamond! per John Thorn: "Green is the color of
spring, of hope, of baseball. In the green fields of the mind, to borrow
Bart Giamatti’s marvelous phrase for the place where hope resides,
baseball players think green thoughts. Why not get that first home run,
or hit .400, or win the pennant?"