038: adding photo website to table
also adding a gitignore? we'll see if that works

also had the showr thought of every time I update the table, adding a
post to instagram? frankly i don't like the extra lift, and that bit of
friciton will make me want to update less. besides, what good is the
instagram, anyway? maybe this is best as an IYKYK thing